switch token access to a function in package
advantage of this is if you ever have to change the implementation of
how tokens are get and set, perhaps as the result of a change to the API
you no longer have to update all definitions.

corresponsing set_token() function makes it easier for users to know how
to set their token.

add associated tests